// autogpt_platform/frontend/src/app/api/__generated__/models/blockInfo.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { BlockInfoInputSchema } from "./blockInfoInputSchema";
import type { BlockInfoOutputSchema } from "./blockInfoOutputSchema";
import type { BlockCost } from "./blockCost";
import type { BlockInfoCategoriesItem } from "./blockInfoCategoriesItem";
import type { BlockInfoContributorsItem } from "./blockInfoContributorsItem";

export interface BlockInfo {
  id: string;
  name: string;
  inputSchema: BlockInfoInputSchema;
  outputSchema: BlockInfoOutputSchema;
  costs: BlockCost[];
  description: string;
  categories: BlockInfoCategoriesItem[];
  contributors: BlockInfoContributorsItem[];
  staticOutput: boolean;
  uiType: string;
}